use in-place functions for result to state/effect transformation
address #20
